ml-service/requirements-dev.txt
# ML Service Development - Requirements for project setup
# Install with: pip install -r requirements.txt

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# ML Libraries (uncomment after initial setup)
# scikit-learn>=1.3.0
# pandas>=2.0.0
# numpy>=1.24.0
# joblib>=1.3.0

# Data visualization (for notebooks)
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Jupyter for development
# jupyter>=1.0.0
# notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional but recommended
# xgboost>=2.0.0  # For advanced models
# lightgbm>=4.0.0  # For advanced models
